+++ src/GlobalPayments.Api/Entities/DocumentUploadData.cs
@@ -15,13 +15,28 @@ namespace GlobalPayments.Api.Entities
         /// </summary>
         public string TransactionReference { get; set; }
         /// <summary>
-        /// The file format of the Document to be uploaded. This value is set automatically by setting the DocumentPath property
+        /// The file format of the Document to be uploaded. This property MUST be set if using the Document property directly, but will be set automatically if using the DocumentPath property
         /// </summary>
-        public string DocType { get; private set; }
+        private string _docType;
+        public string DocType
+        {
+            get { return _docType; }
+            set
+            {
+                if (_validDocTypes.Contains(value)) {
+                    _docType = value;
+                }
+                else {
+                    throw new Exception("The provided file type is not supported.");
+                }
+            }
+        }
         /// <summary>
-        /// The document data in base64 format. This value is set automatically by setting the DocumentPath property
+        /// The document data in base64 format.
+        /// This property can be assigned to directly (the DocType property must also be provided a value) or
+        /// This property will be set automatically by setting the DocumentPath property
         /// </summary>
-        public string Document { get; private set; }
+        public string Document { get; set; }
         /// <summary>
         /// The type of document you've been asked to provide by ProPay's Risk team. Valid values are:
         /// Verification, FraudHolds, Underwriting, RetrievalRequest
@@ -31,21 +46,17 @@ namespace GlobalPayments.Api.Entities
             set
             {
                 var docPath = value;
-                if (docPath != null)
-                {
+                if (docPath != null) {
                         var docType = docPath.Substring(docPath.LastIndexOf('.') + 1);
-                        if (_validDocTypes.Contains(docType))
-                        {
+                        if (_validDocTypes.Contains(docType)) {
                             DocType = docType;
                             Document = Convert.ToBase64String(System.IO.File.ReadAllBytes(docPath));
                         }
-                        else
-                        {
+                        else {
                             throw new Exception("The document provided is not a valid file type.");
                         }
                     }
-                    else
-                    {
+                    else {
                         throw new Exception("DocumentPath has not been set");
                     }
             }